feat: adding responsive layout
