+++ src/app/components/GameNotStarted/GameNotStarted.module.scss
