+++ src/app/components/GameNotStarted/index.tsx
