+++ src/app/components/GameStarted/GameStarted.tsx
@@ -1,19 +1,23 @@
 import styles from "./GameStarted.module.scss"
 import Button from "../Button";
-import React from "react";
+import React, {useContext} from "react";
 import { formatCurrency } from "../../../utils/formatUtils";
+import {AppContext} from "../../context/AppContext";
+import IconTooltip from "../IconTooltip";
 
 interface GameStartedProps {
     betValue: number
 }
 
 const GameStarted = ({ betValue }: GameStartedProps) => {
+    const { isMobile } = useContext(AppContext)
+
     return (
         <div className={styles.fullpage}>
             <div className={styles.sideContainer}>
-                <img src="/hand-holding-card.webp" alt="Hand holding a card" className={styles.logo}/>
+                {!isMobile && <img src="/hand-holding-card.webp" alt="Hand holding a card" className={styles.logo}/>}
                 <h1 className={styles.title}>BLACKJACK</h1>
-                <p className={styles.text}>Chegue o mais perto possível <br/> de 21 sem ultrapassar!</p>
+                {!isMobile && < p className={styles.text}>Chegue o mais perto possível <br/> de 21 sem ultrapassar!</p>}
                 <p className={styles.text}>Você tem {formatCurrency(betValue, true)} apostados nessa rodada</p>
             </div>
             <div className={styles.container}>
@@ -21,8 +25,11 @@ const GameStarted = ({ betValue }: GameStartedProps) => {
                 <h3>Card</h3>
                 <h1 className={styles.title}>Sua mão</h1>
                 <h3>Card</h3>
-                <Button label="Comprar carta" onClick={() => true}/>
-                <Button label="Permanecer" onClick={() => true}/>
+                <div>
+                    <Button label="Comprar carta" onClick={() => true}/>
+                    <Button label="Permanecer" onClick={() => true}/>
+                    {isMobile && <IconTooltip />}
+                </div>
             </div>
         </div>
     )
