+++ src/app/components/GameStarted/index.tsx
