+++ src/app/components/IconTooltip/index.tsx
@@ -1,0 +1,3 @@
+import IconTooltip from "./IconTooltip"
+
+export default IconTooltip
