+++ src/app/components/MainPage.module.scss
