+++ src/app/components/MoneyInput/MoneyInput.module.scss
