+++ src/app/context/AppContext.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface AppContextProps {
+    isMobile: boolean;
+}
+
+export const AppContext = React.createContext<AppContextProps>({
+    isMobile: false,
+})
