+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import { MainPage } from './App/MainPage';
+import { MainPage } from './app/components/MainPage';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
