+++ src/mixins.scss
@@ -1,0 +1,14 @@
+@mixin respond($breakpoint) {
+  @if $breakpoint == phone {
+    @media (max-width: 600px) { @content; }
+  }
+  @if $breakpoint == tablet-portrait {
+    @media (max-width: 900px) { @content; }
+  }
+  @if $breakpoint == desktop-medium {
+    @media (max-width: 1200px) { @content; }
+  }
+  @if $breakpoint == desktop-large {
+    @media (max-width: 1800px) { @content; }
+  }
+}
